backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
transformers
huggingface-hub
requests
